Rename generic 'util' file
